requirements.txt
python-telegram-bot>=20.0
python-nmap
apscheduler>=4.0.0
